backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic
pillow
torch
timm
email-validator